add styles

+++ src/app/remove-dog.component.ts
@@ -3,8 +3,7 @@ import { DogsService } from './dogs-races.service';
 
 @Component({
     selector: 'app-remove-dog',
-    templateUrl: './remove-dog.component.html',
-    // styleUrls: ['./name.component.scss']
+    templateUrl: './remove-dog.component.html'
 })
 export class RemoveDogComponent {
     @Input() dogId: number;
@@ -12,7 +11,7 @@ export class RemoveDogComponent {
     constructor(private dogsService: DogsService) { }
 
     RemoveDog(){
-        if(this.dogId){
+        if(this.dogId && confirm("Are you sure to delete")){
             this.dogsService.removeDog(this.dogId)
             .subscribe(
                 res => {
